% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol_relapse}
\alias{alcohol_relapse}
\title{Weeks to alcohol relapse after treatment}
\format{
A person-level data frame with 89 rows and
3 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{weeks}}{Number of weeks to first "heavy drinking" day}
\item{\code{censor}}{Censoring status.}
}
}
\source{
Cooney, N. L., Kadden, R. M., Litt, M. D., & Getter, H. (1991). Matching
alcoholics to coping skills or interactional therapies: Two-year follow-up
results. Journal of Consulting and Clinical Psychology, 59, 598–601.
\url{https://doi.org/10.1037/0022-006X.59.4.598}
}
\usage{
alcohol_relapse
}
\description{
A subset of data from Cooney and colleagues (1991) measuring weeks to first
"heavy drinking" day in a sample of 89 recently treated alcoholics.
Individuals were followed for up to two years (around 104.286 weeks) or until
they relapsed.
}
\keyword{datasets}
