% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol_use_1}
\alias{alcohol_use_1}
\title{Adolescent and peer alcohol use}
\format{
A person-period data frame with 246 rows and
7 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{age}}{in years.}
\item{\code{child_of_alcoholic}}{Binary indicator for whether the adolescent is a child of an alcoholic parent.}
\item{\code{male}}{Binary indicator for whether the adolescent is a male.}
\item{\code{alcohol_use}}{Square root of the summed scores of four eight-point items measuring frequency of alcohol use.}
\item{\code{peer_alcohol_use}}{Square root of the summed scores of two six-point items measuring frequency of peer alcohol use.}
}
}
\source{
Curran, P. J., Stice, E., & Chassin, L. (1997). The relation between
adolescent and peer alcohol use: A longitudinal random coefficients model.
Journal of Consulting and Clinical Psychology, 65, 130–140.
\url{https://doi.org/10.1037//0022-006x.65.1.130}
}
\usage{
alcohol_use_1
}
\description{
A subset of data from Curran, Stice, and Chassin (1997) measuring the
relation between changes in alcohol use and changes in peer alcohol use over
a 3-year period in a community-based sample of Hispanic and Caucasian
adolescents.
}
\keyword{datasets}
