% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol_use_2}
\alias{alcohol_use_2}
\title{Adolescent alcohol consumption and peer pressure}
\format{
A person-period data frame with 3366 rows and
5 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{time}}{Time of measurement.}
\item{\code{female}}{Binary indicator for whether the adolescent is a female.}
\item{\code{alcohol_use}}{Natural logarithm of the averaged scores of three six-point items measuring frequency of beer, wine, and liquor consumption, respectively.}
\item{\code{peer_pressure}}{Natural logarithm of a six-point item measuring frequency friends offered alcoholic drinks during the past month.}
}
}
\source{
Barnes, G. M., Farrell, M. P., & Banerjee, S. (1994). Family influences on
alcohol abuse and other problem behaviors among black and white adolescents
in a general population sample. Journal of Research on Adolescence, 4,
183–201. \url{https://doi.org/10.1207/s15327795jra0402_2}
}
\usage{
alcohol_use_2
}
\description{
Data from Barnes, Farrell, and Banerjee (1994) measuring the relation between
changes in alcohol use and changes in peer pressure to use alcohol in a
sample of 1122 Black and White adolescents tracked from the beginning of
seventh grade through the end of eighth grade.
}
\note{
Barnes, Farrell, and Banerjee (1994) report a sample of only 699 adolescents;
however, they note that this was an ongoing longitudinal study, which likely
explains the sample size discrepancy with the data used by Singer and Willett
(2003).
}
\keyword{datasets}
