% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antidepressants}
\alias{antidepressants}
\title{Antidepressant medication and positive mood}
\format{
A person-period data frame with 1242 rows and
8 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{wave}}{Wave of measurement.}
\item{\code{day}}{Day of measurement.}
\item{\code{reading}}{Time of day of measurement.}
\item{\code{time_of_day}}{Time of day of measurement expressed numerically (0 for morning readings; 0.33 for afternoon readings; 0.67 for evening readings).}
\item{\code{time}}{Time of measurement expressed as a combination of \code{day} and \code{time_of_day}.}
\item{\code{positive_mood}}{Positive mood score.}
\item{\code{treatment}}{Treatment condition (placebo pills = 0, antidepressant pills = 1)}
}
}
\source{
Tomarken, A. J., Shelton, R. C., Elkins, L., & Anderson, T. (1997). Sleep
deprivation and anti-depressant medication: Unique effects on positive and
negative affect. Poster session presented at the 9th annual meeting of the
American Psychological Society, Washington, DC.
}
\usage{
antidepressants
}
\description{
A subset of data from Tomarken, Shelton, Elkins, and Anderson (1997)
measuring the relation between changes in positive mood and supplemental
antidepressant medication over the course of a week in a sample of 73 men and
women already receiving nonpharmacological therapy for depression.
}
\keyword{datasets}
