% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berkeley}
\alias{berkeley}
\title{Berkeley Growth Study}
\format{
A person-period data frame with 18 rows and
2 columns:

\describe{
\item{\code{age}}{Age of the girl in years.}
\item{\code{iq}}{IQ score.}
}
}
\source{
Bayley, N. (1935). The development of motor abilities during the first
three years. Monographs of the Society for Research in Child Development, 1.
}
\usage{
berkeley
}
\description{
A subset of data from the Berkeley Growth Study measuring changes in IQ of a
single girl followed from childhood into older adulthood (Bayley, 1935).
}
\keyword{datasets}
