% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cocaine_relapse_1}
\alias{cocaine_relapse_1}
\title{Weeks to cocaine relapse after treatment}
\format{
A person-level data frame with 104 rows and
4 columns:

\describe{
\item{\code{id}}{In-patient ID.}
\item{\code{weeks}}{The number of weeks between an in-patient's release and relapse to cocaine use.}
\item{\code{censor}}{Censoring status.}
\item{\code{needle}}{Binary indicator for whether cocaine was ever used intravenously.}
}
}
\source{
Hall, S. M., Havassy, B. E., & Wasserman, D. A. (1990). Commitment to
abstinence and acute stress in relapse to alcohol, opiates, and nicotine.
Journal of Consulting and Clinical Psychology, 58, 175–181.
\url{https://doi.org/10.1037//0022-006x.58.2.175}
}
\usage{
cocaine_relapse_1
}
\description{
A subset of data from Hall, Havassy, and Wasserman (1990) measuring the
number of weeks of relapse to cocaine use in a sample of 104 former addicts
released from an in-patient treatment program. In-patients were followed for
up to 12 weeks or until they used cocaine for 7 consecutive days.
}
\keyword{datasets}
