% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cocaine_relapse_2}
\alias{cocaine_relapse_2}
\title{Days to cocaine relapse after abstinence}
\format{
A person-period data frame with 1248 rows and
7 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{days}}{Number of days until relapse to cocaine use or censoring. Relapse was defined as 4 or more days of cocaine use during the week preceding an interview. Study dropouts and lost participants were coded as relapsing to cocaine use, with the number of days until relapse coded as occurring the week after the last follow-up interview attended.}
\item{\code{censor}}{Censoring status (0 = relapsed, 1 = censored).}
\item{\code{needle}}{Binary indicator for whether cocaine was ever used intravenously.}
\item{\code{base_mood}}{Total score on the positive mood subscales (Activity and Happiness) of the Mood Questionnaire (Ryman, Biersner, & LaRocco, 1974), taken at an intake interview during the last week of treatment. Each item used a five point Likert score (ranging from 0 = not at all, to 4 = extremely).}
\item{\code{followup}}{Week of follow-up interview.}
\item{\code{mood}}{Total score on the positive mood subscales (Activity and Happiness) of the Mood Questionnaire (Ryman, Biersner, & LaRocco, 1974), taken during follow-up interviews each week post-treatment. Each item used a five point Likert score (ranging from 0 = not at all, to 4 = extremely).}
}
}
\source{
Hall, S. M., Havassy, B. E., & Wasserman, D. A. (1990). Commitment to
abstinence and acute stress in relapse to alcohol, opiates, and nicotine.
Journal of Consulting and Clinical Psychology, 58, 175–181.
\url{https://doi.org/10.1037//0022-006x.58.2.175}
}
\usage{
cocaine_relapse_2
}
\description{
A subset of unpublished data from Hall, Havassy, and Wasserman (1990)
measuring the relation between the number of days until relapse to cocaine
use and several predictors that might be associated with relapse in a sample
of 104 newly abstinent cocaine users who recently completed an
abstinence-oriented treatment program. Former cocaine users were followed for
up to 12 weeks post-treatment or until they used cocaine for 7 consecutive
days. Self-reported abstinence was confirmed at each interview by the absence
of cocaine in urine specimens.
}
\note{
Hall, Havassy, and Wasserman (1990) measured time to relapse in weeks, not
days; however, to use these data to illustrate imputation strategies, Singer
and Willett (2003) converted the weekly relapse information into days, then
jittered these event times, effectively converting them from discrete-time to
continuous-time. Additionally, Hall, Havassy, and Wasserman (1990) do not
report following cocaine users in their study, thus, this appears to be
unpublished data.
}
\references{
Ryman, D. H., Biersner, R. J., & La Rocco, J. M. (1974). Reliabilities and
validities of the Mood Questionnaire. Psychological Reports, 35, 479-484.
\url{https://doi.org/10.2466/pr0.1974.35.1.479}
}
\keyword{datasets}
