% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cognitive_growth}
\alias{cognitive_growth}
\title{Cognitive growth in children}
\format{
A person-period data frame with 445 rows and
4 columns:

\describe{
\item{\code{id}}{Child ID.}
\item{\code{game}}{Game number. Each child played up to a maximum of 27 games.}
\item{\code{nmoves}}{The number of moves completed before making a catastrophic error.}
\item{\code{read}}{Score on an unnamed standardized reading test.}
}
}
\source{
Tivnan, T. (1980). Improvements in performance on cognitive tasks: The
acquisition of new skills by elementary school children. Unpublished doctoral
dissertation. Harvard University, Graduate School of Education.
}
\usage{
cognitive_growth
}
\description{
Data from Tivnan (1980) measuring changes in cognitive growth over a
three-week period in a sample of 17 firstand second-graders.
}
\keyword{datasets}
