% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congresswomen}
\alias{congresswomen}
\title{House of Representatives tenure}
\format{
A person-level data frame with 168 rows and
5 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{name}}{Representative name.}
\item{\code{terms}}{Number of terms in office.}
\item{\code{censor}}{Censoring status.}
\item{\code{democrat}}{Party affiliation.}
}
}
\usage{
congresswomen
}
\description{
Data measuring how long all 168 women who were elected to the U.S. House of
Representatives between 1919 and 1996 remained in office. Representatives
were followed for up to eight terms or until 1998.
}
\keyword{datasets}
