% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depression_unemployment}
\alias{depression_unemployment}
\title{Unemployment and depression}
\format{
A person-period data frame with 674 rows and
5 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{interview}}{Time of interview.}
\item{\code{months}}{Months since job loss.}
\item{\code{depression}}{Center for Epidemiologic Studies' Depression (CES-D) scale score (Radloff, 1977)}
\item{\code{unemployed}}{Binary indicator for whether the participant was unemployed at time of interview.}
}
}
\source{
Ginexi, E. M., Howe, G. W., & Caplan, R. D. (2000). Depression and control
beliefs in relation to reemployment: What are the directions of effect?
Journal of Occupational Health Psychology, 5, 323–336.
\url{https://doi.org/10.1037//1076-8998.5.3.323}
}
\usage{
depression_unemployment
}
\description{
A subset of data from Ginexi and colleagues (2000) measuring changes in
depressive symptoms after job loss in a sample of 254 recently unemployed men
and women. Interviews were conducted in three waves at around 1, 5, and 12
months after job loss.
}
\references{
Radloff, L. S. (1977). The CES-D scale: A self report major depressive
disorder scale for research in the general population. Applied Psychological
Measurement, 1, 385–401. \url{https://doi.org/10.1177/014662167700100306}
}
\keyword{datasets}
