% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deviant_tolerance_pp}
\alias{deviant_tolerance_pp}
\alias{deviant_tolerance}
\alias{deviant_tolerance_pl}
\title{Adolescent tolerance of deviant behaviour}
\format{
\subsection{\code{deviant_tolerance_pp}}{

A person-period data frame with 80 rows and
5 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{age}}{Adolescent age in years.}
\item{\code{tolerance}}{Average score across a 9-item scale assessing attitudes favourable to deviant behaviour. Each item used a four point scale (1 = very wrong, 2 = wrong, 3 = a little bit wrong, 4 = not wrong at all).}
\item{\code{male}}{Binary indicator for whether the adolescent is a male.}
\item{\code{exposure}}{Average score across a 9-item scale assessing level of exposure to deviant peers. Each item used a five point Likert score (ranging from 0 = none, to 4 = all).}
}
}

\subsection{\code{deviant_tolerance_pl}}{

A person-level data frame with 16 rows and
8 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{tolerance_11}, \code{tolerance_12}, \code{tolerance_13}, \code{tolerance_14}, \code{tolerance_15},}{Average score across a 9-item scale assessing attitudes favourable to deviant behaviour at ages 11, 12, 13, 14, and 15. Each item used a four point scale (1 = very wrong, 2 = wrong, 3 = a little bit wrong, 4 = not wrong at all).}
\item{\code{male}}{Binary indicator for whether the adolescent is a male.}
\item{\code{exposure}}{Average score across a 9-item scale assessing level of exposure to deviant peers. Each item used a five point Likert score (ranging from 0 = none, to 4 = all).}
}
}
}
\source{
Raudenbush, S. W., & Chan, W. S. (1992). Growth curve analysis in
accelerated longitudinal designs. Journal of Research in Crime and
Delinquency, 29, 387–411. \url{https://doi.org/10.1177/0022427892029004001}
}
\usage{
deviant_tolerance_pp

deviant_tolerance_pl
}
\description{
A subset of data from the National Youth Survey (NYS) measuring tolerance of
deviant behaviour in adolescents over time (Raudenbush & Chan, 1992).
}
\note{
Raudenbush and Chan (1992) comment that \code{exposure} was a time-varying
predictor in the original study; however, Singer and Willett (2003) provide
\code{exposure} as a time-invariant predictor.
}
\keyword{datasets}
