% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{externalizing_behaviour}
\alias{externalizing_behaviour}
\title{Externalizing behaviour in children}
\format{
A person-period data frame with 270 rows and
5 columns:

\describe{
\item{\code{id}}{Child ID.}
\item{\code{time}}{Time of measurement.}
\item{\code{externalizing_behaviour}}{Sum of scores on Achenbach's (1991) Child Behavior Checklist. Scores range from 0 to 68}
\item{\code{female}}{Binary indicator for whether the adolescent is a female.}
\item{\code{grade}}{Grade year.}
}
}
\source{
Keiley, M. K., Bates, J. E., Dodge, K. A., & Pettit, G. S. (2000). A
cross-domain growth analysis: Externalizing and internalizing behavior during
8 years of childhood. Journal of Abnormal Child Psychology, 28, 161–179.
\url{https://doi.org/10.1023\%2Fa\%3A1005122814723}
}
\usage{
externalizing_behaviour
}
\description{
A subset of data from Keiley, Bates, Dodge, and Pettit (2000) measuring
changes in externalizing behaviour in a sample of 45 children tracked from
first through sixth grade.
}
\references{
Achenbach, T. M. (1991). Manual for the Child Behavior Checklist 4–18 and
1991 Profile. Burlington, VT: University of Vermont Press.
}
\keyword{datasets}
