% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{first_arrest}
\alias{first_arrest}
\title{Age of first juvenile arrest}
\format{
A person-period data frame with 15834 rows and
19 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{time}}{Age of first juvenile arrest.}
\item{\code{censor}}{Censoring status.}
\item{\code{abused}}{Binary indicator for whether the adolescent was abused.}
\item{\code{black}}{Binary indicator for whether the adolescent is black.}
\item{\code{abused_black}}{Binary indicator for whether the adolescent was abused and is black.}
\item{\code{d8-d18}}{Discrete time indicators for each age.}
\item{\code{period}}{Age each record corresponds to.}
\item{\code{event}}{Binary indicator for whether the adolescent was arrested.}
}
}
\source{
Keiley, M. K., & Martin, N. C. (2002). Child abuse, neglect, and juvenile
delinquency: How “new” statistical approaches can inform our understanding of
“old” questions—a reanalysis of Widon, 1989. Manuscript submitted for
publication.
}
\usage{
first_arrest
}
\description{
Data from Keiley and Martin (2002) measuring the effect of child abuse on the
risk of first juvenile arrest in a sample of 1553 adolescents aged 8 to 18.
Adolescents were followed up to age 18 or until they were arrested.
}
\keyword{datasets}
