% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{first_cocaine}
\alias{first_cocaine}
\title{Age of first cocaine use}
\format{
A person-level data frame with 1658 rows and
15 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{used_cocaine_age}}{Age of first cocaine use.}
\item{\code{censor}}{Censoring status.}
\item{birth_year}{}
\item{\code{early_marijuana_use}}{Binary indicator for whether marijuana was used before age 17.}
\item{\code{used_marijuana}}{Binary indicator for whether the participant used marijuana during the study period.}
\item{\code{used_marijuana_age}}{Age the participant first used marijuana.}
\item{\code{sold_marijuana}}{Binary indicator for whether the participant sold marijuana during the study period.}
\item{\code{sold_marijuana_age}}{Age the participant first sold marijuana.}
\item{\code{early_drug_use}}{Binary indicator for whether other drugs were used before age 17.}
\item{\code{used_drugs}}{Binary indicator for whether the participant used other drugs during the study period.}
\item{\code{used_drugs_age}}{Age the participant first used other drugs.}
\item{\code{sold_drugs}}{Binary indicator for whether the participant sold other drugs during the study period.}
\item{\code{sold_drugs_age}}{Age the participant first sold other drugs.}
\item{\code{rural}}{Binary indicator for whether the participant lived in a rural area.}
}
}
\source{
Burton, R. P. D., Johnson, R. J., Ritter, C., & Clayton. R. R. (1996). The
effects of role socialization on the initiation of cocaine use: An event
history analysis from adolescence into middle adulthood. Journal of Health
and Social Behavior, 37, 75–90. \url{https://doi.org/10.2307/2137232}
}
\usage{
first_cocaine
}
\description{
Data from Burton and colleagues (1996) measuring the relation between age of
first cocaine use and drug-use history in a random sample of 1658 white
American men. Age of first cocaine use and drug-use history was determined
through two interviews eleven years apart (1974 and 1985).
}
\keyword{datasets}
