% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{first_depression_2}
\alias{first_depression_2}
\title{Age of first depression}
\format{
A person-level data frame with 2974 rows and
3 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{age}}{in years.}
\item{\code{censor}}{Censoring status.}
}
}
\source{
Sorenson, S. B., Rutter, C. M., & Aneshensel, C. S. (1991). Depression in
the community: An investigation into age of onset. Journal of Consulting and
Clinical Psychology, 59, 541546. \url{https://doi.org/10.1037/0022-006X.59.4.541}
}
\usage{
first_depression_2
}
\description{
Data from Sorenson, Rutter, and Aneshensel (1991) measuring age of first
depressive episode in a sample of 2974 adults. Age of first depressive
episode was measured by asking respondents whether and, if so, at what age
they first experienced a depressive episode.
}
\keyword{datasets}
