% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health_workers}
\alias{health_workers}
\title{Length of health worker employment}
\format{
A person-level data frame with 2074 rows and
3 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{weeks}}{Number of weeks to termination of employment.}
\item{\code{censor}}{Censoring status.}
}
}
\source{
Singer, J. D., Davidson, S., Graham, S., & Davidson, H. S. (1998).
Physician retention in community and migrant health centers: Who stays and
for how long? Medical Care, 38, 11981213.
\url{https://doi.org/10.1097/00005650-199808000-00008}
}
\usage{
health_workers
}
\description{
A subset of data from Singer and colleagues (1998) measuring length of
employment in a sample of 2074 health care workers hired by community and
migrant health centres. Health care workers were followed for up to 33 months
or until termination of employment.
}
\keyword{datasets}
