% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{judges}
\alias{judges}
\title{Supreme Court justice tenure}
\format{
A person-level data frame with 109 rows and
7 columns:

\describe{
\item{\code{id}}{Justice ID.}
\item{\code{tenure}}{Time to retirement or death in years.}
\item{\code{dead}}{Binary indicator for whether the justice died.}
\item{\code{retired}}{Binary indicator for whether the justice retired.}
\item{\code{left_appointment}}{Binary indicator for whether the justice left their appointment.}
\item{\code{appointment_age}}{Age at time of appointment.}
\item{\code{appointment_year}}{Year of appointment.}
}
}
\source{
Zorn, C. J., & van Winkle, S. R. (2000). A competing risks model of Supreme
Court vacancies, 1780–1992. Political Behavior, 22, 145–166.
}
\usage{
judges
}
\description{
Data from Zorn and Van Winkle (2000) how long all 107 justices appointed to
the U.S. Supreme Court between 1789 and 1980 remained in their position.
}
\keyword{datasets}
