% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monkeys}
\alias{monkeys}
\title{Piagetian monkeys}
\format{
A person-level data frame with 123 rows and
7 columns:

\describe{
\item{\code{id}}{Monkey ID.}
\item{\code{sessions}}{Number of sessions the monkey completed until demonstrating object recognition.}
\item{\code{initial_age}}{Age at initial testing in days.}
\item{\code{end_age}}{Age at end of testing in days.}
\item{\code{censor}}{Censoring status.}
\item{\code{birth_weight}}{Decile equivalent of the monkey's birth weight in comparison to colony-wide sex-specific standards.}
\item{\code{female}}{Binary indicator for whether the adolescent is a female.}
}
}
\source{
Ha, J. C., Kimpo, C. L., & Sackett, G. P. (1997). Multiple-spell,
discrete-time survival analysis of developmental data: Object concept in
pigtailed macaques. Developmental Psychology, 33, 1054–1059.
\url{https://doi.org/10.1037//0012-1649.33.6.1054}
}
\usage{
monkeys
}
\description{
Data from Ha, Kimpo, and Sackett (1997) measuring age of first demonstration
of object recognition in a sample of 123 pigtailed macaques. Monkeys were
followed for up to 37 days or until they demonstrated the classic Piagetian
stage of development known as object recognition.
}
\keyword{datasets}
