% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opposites_naming}
\alias{opposites_naming}
\title{Opposites naming Task}
\format{
A person-period data frame with 140 rows and
5 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{wave}}{Wave of measurement.}
\item{\code{time}}{Wave of measurement centred at time 0.}
\item{\code{opposites_naming_score}}{Score on the "opposites naming" task.}
\item{\code{baseline_cognitive_score}}{Baseline score on a standardized instrument assessing general cognitive skill.}
}
}
\source{
Willett, J. B. (1988). Questions and answers in the measurement of change.
In E. Rothkopf (Ed.), Review of research in education (1988–89) (pp.
345–422). Washington, DC: American Educational Research Association.
\url{https://doi.org/10.3102/0091732X015001345}
}
\usage{
opposites_naming
}
\description{
Artificial data created by Willett (1988) measuring changes in performance on
a hypothetical "opposites naming" task over a four week period in a sample of
35 people.
}
\keyword{datasets}
