% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physicians}
\alias{physicians}
\title{Physician career history}
\format{
A person-level data frame with 812 rows and
8 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{start_date}}{Date of hire.}
\item{\code{end_date}}{Date of departure.}
\item{\code{entry}}{Number of years since hire the physician worked before entering the measurement window.}
\item{\code{exit}}{Number of years the physician worked until departure.}
\item{\code{censor}}{Censoring status.}
\item{\code{part_time}}{Binary indicator for whether the physician worked part time.}
\item{\code{age}}{Age at time of hire.}
}
}
\source{
Singer, J. D., Davidson, S., Graham, S., & Davidson, H. S. (1998).
Physician retention in community and migrant health centers: Who stays and
for how long? Medical Care, 38, 11981213.
\url{https://doi.org/10.1097/00005650-199808000-00008}
}
\usage{
physicians
}
\description{
A subset of data from Singer and colleagues (1998) measuring length of
employment in a sample of 812 physicians hired by community and migrant
health centres. Physicians were followed for up to 33 months or until
termination of employment. The measurement window began on January 1, 1990,
and ended on September 30, 1992.
}
\keyword{datasets}
