% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psychiatric_discharge}
\alias{psychiatric_discharge}
\title{Days to psychiatric hospital discharge}
\format{
A person-level data frame with 174 rows and
4 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{days}}{Number of days to discharge.}
\item{\code{censor}}{Censoring status.}
\item{\code{treatment_plan}}{Binary indicator for whether the patient had a traditional coverage plan (0) or an innovative coverage plan (1).}
}
}
\source{
Foster, E. M. (2000). Does the continuum of care reduce inpatient length of
stay? Evaluation and Program Planning, 23, 53–65.
\url{https://doi.org/10.1016/S0149-7189(99)00037-3}
}
\usage{
psychiatric_discharge
}
\description{
A subset of data from Foster (2000) measuring the relation between the number
of days to discharge from a psychiatric hospital and type of treatment plan
in a sample of 174 adolescents with emotional and behavioural problems.
}
\keyword{datasets}
