% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reading_scores}
\alias{reading_scores}
\title{Peabody Individual Achievement Test reading scores}
\format{
A person-period data frame with 267 rows and
5 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{wave}}{Wave of measurement.}
\item{\code{age_group}}{Expected age on each measurement occasion.}
\item{\code{age}}{Age in years at time of measurement.}
\item{\code{reading_score}}{Reading score on the reading subtest of the Peabody Individual Achievement Test (PIAT).}
}
}
\source{
US Bureau of Labor and Statistics. National Longitudinal Survey of Youth
(Children of the NLSY). \url{https://www.bls.gov/nls/nlsy79-children.htm}
}
\usage{
reading_scores
}
\description{
A subset of data from the Children of the National Longitudinal Study of
Youth measuring changes in the reading subtest of the Peabody Individual
Achievement Test (PIAT) in a sample of 89 African-American children across
three waves around the ages of 6, 8, and 10.
}
\keyword{datasets}
