% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rearrest}
\alias{rearrest}
\title{Days to inmate recidivism}
\format{
A person-level data frame with 194 rows and
7 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{days}}{Number of days to rearrest.}
\item{\code{months}}{Number of months to rearrest, on the scale of an "average" month (with 30.4375 days).}
\item{\code{censor}}{Censoring status.}
\item{\code{personal}}{Committed a person-related crime}
\item{\code{property}}{Binary indicator for whether the inmate committed a property crime.}
\item{\code{age}}{Centred age at time or release.}
}
}
\source{
Henning, K. R., & Frueh, B. C. (1996). Cognitive-behavioral treatment of
incarcerated offenders: An evaluation of the Vermont Department of
Corrections' cognitive self-change program. Criminal Justice and Behavior,
23, 523–541. \url{https://doi.org/10.1177/0093854896023004001}
}
\usage{
rearrest
}
\description{
Data from Henning and Frueh (1996) measuring measuring days to rearrest in
a sample of 194 inmates recently released from a medium security prison. Inmates
were followed for up to three years or until they were rearrested.
}
\keyword{datasets}
