% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suicide_ideation}
\alias{suicide_ideation}
\title{Age of first suicide ideation}
\format{
A person-level data frame with 391 rows and
4 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{time}}{Reported age of first suicide ideation.}
\item{\code{censor}}{Censoring status.}
\item{\code{age}}{Participant age at the time of the survey.}
}
}
\source{
Bolger, N., Downey, G., Walker, E., & Steininger, P. (1989). The onset of
suicide ideation in childhood and adolescence. Journal of Youth and
Adolescence, 18, 175–189. \url{https://doi.org/10.1007/BF02138799}
}
\usage{
suicide_ideation
}
\description{
A subset of data from Bolger and colleagues (1989) measuring age of first
suicide ideation in a sample of 391 undergraduate students aged 16 to 22.
Age of first suicide ideation was measured with a two-item survey asking
respondents "Have you ever thought of committing suicide?" and if so, "At
what age did the thought first occur to you?"
}
\keyword{datasets}
