% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teachers}
\alias{teachers}
\title{Years to special education teacher turnover}
\format{
A person-level data frame with 3941 rows and
3 columns:

\describe{
\item{\code{id}}{Teacher ID.}
\item{\code{years}}{The number of years between a teacher's dates of hire and departure from the Michigan public schools.}
\item{\code{censor}}{Censoring status.}
}
}
\source{
Singer, J. D. (1992). Are special educators' careers special? Results from
a 13-Year Longitudinal Study. Exceptional Children, 59, 262–279.
\url{https://doi.org/10.1177/001440299305900309}
}
\usage{
teachers
}
\description{
A subset of data from Singer (1993) measuring how many years 3941 newly hired
special educators in Michigan stayed in teaching between 1972 and 1978.
Teachers were followed for up to 13 years or until they stopped teaching in
the state.
}
\keyword{datasets}
