% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenure}
\alias{tenure}
\title{Years to academic tenure}
\format{
A person-level data frame with 260 rows and
3 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{years}}{Number of years to receiving tenure.}
\item{\code{censor}}{Censoring status.}
}
}
\source{
Gamse, B. C., & Conger, D. (1997). An evaluation of the Spencer
post-doctoral dissertation program. Cambridge, MA: Abt Associates.
}
\usage{
tenure
}
\description{
Data from Gamse and Conger (1997) measuring the number of years to receiving
tenure in a sample of 260 semifinalists and fellowship recipients in the
National Academy of Education–Spencer Foundation PostDoctoral Fellowship
Program who took an academic job after earning a doctorate. Academics were
followed for up to nine years or until they received tenure.
}
\keyword{datasets}
