% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cocaine_relapse_2}
\alias{cocaine_relapse_2}
\title{Days to cocaine relapse after abstinence}
\format{
A person-period data frame with 1248 rows and
7 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{days}}{Number of days of relapse to cocaine use.}
\item{\code{censor}}{Censoring status.}
\item{\code{needle}}{Binary indicator for whether cocaine was ever used intravenously.}
\item{\code{base_mood}}{Positive mood score on a standardized question at an intake interview taken during the last week of treatment.}
\item{\code{followup}}{Week of follow-up interview.}
\item{\code{mood}}{Positive mood score on a standardized question.}
}
}
\source{
Hall, S. M., Havassy, B. E., & Wasserman, D. A. (1990). Commitment to
abstinence and acute stress in relapse to alcohol, opiates, and nicotine.
Journal of Consulting and Clinical Psychology, 58, 175–181.
\url{https://doi.org/10.1037//0022-006x.58.2.175}
}
\usage{
cocaine_relapse_2
}
\description{
Data from Hall and colleagues (1990) measuring the relation between the
number of days of relapse to cocaine use and several predictors that might
be associated relapse in a sample of 104 newly abstinent cocaine users.
Former addicts were followed for up to 12 weeks or until relapse.
}
\keyword{datasets}
