% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{first_depression_1}
\alias{first_depression_1}
\title{Age of first depression}
\format{
A person-period data frame with 36997 rows and
11 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{onset}}{Age of first depressive episode.}
\item{\code{censor}}{Censoring status.}
\item{\code{interview_age}}{Age at time of interview.}
\item{\code{female}}{Binary indicator for whether the adult is a female.}
\item{\code{nsibs}}{Number of siblings.}
\item{\code{bigfamily}}{Binary indicator for whether the adult has five or more siblings.}
\item{\code{period}}{Age each record corresponds to.}
\item{\code{depress}}{Binary indicator for whether the adult experienced a depressive episode.}
\item{\code{pd}}{Binary indicator for whether the adult's parents divorced at this or any previous age.}
}
}
\source{
Wheaton, B., Roszell, P., & Hall, K. (1997). The impact of twenty childhood
and adult traumatic stressors on the risk of psychiatric disorder. In I. H.
Gotlib & B. Wheaton (Eds.), Stress and adversity over the life course:
Trajectories and turning points (pp. 50–72). New York: Cambridge University
Press. \url{https://doi.org/10.1017/CBO9780511527623.003}
}
\usage{
first_depression_1
}
\description{
A subset of data from Wheaton, Rozell, and Hall (1997) measuring the relation
between age of first depressive episode and several childhood and adult
traumatic stressors in a random sample of 1393 adults living in metropolitan
Toronto, Ontario. Age of first depressive episode and traumatic stressors was
determined through a structured interview.
}
\keyword{datasets}
