% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{math_dropout}
\alias{math_dropout}
\title{Math course History}
\format{
A person-period data frame with 9558 rows and
17 columns:

\describe{
\item{\code{id}}{Participant ID.}
\item{\code{last_term}}{The term a student stopped enrolling in mathematics courses.}
\item{\code{woman}}{Binary indicators for whether the student identified as a woman.}
\item{\code{censor}}{Censoring status.}
\item{\code{hs11-hs12}}{Discrete time indicators for each term of high school.}
\item{\code{coll1-coll3}}{Discrete time indicators for each term of college.}
\item{\code{term}}{Term each record corresponds to.}
\item{\code{event}}{Binary indicator for whether the student stopped enrolling in mathematics courses at a given term.}
}
}
\source{
Graham, S. E. (1997). The exodus from mathematics: When and why?
Unpublished doctoral dissertation. Harvard University, Graduate School of
Education.
}
\usage{
math_dropout
}
\description{
Data from Graham (1997) measuring the relation between mathematics
course-taking and gender identity in a sample of 3790 tenth grade high school
students. Students were followed for up to 5 terms (eleventh grade, twelfth
grade, and the first three semesters of college) or until they stopped
enrolling in mathematics courses.
}
\keyword{datasets}
